+++ src/test/scala/edu/ie3/simona/test/common/input/NodeInputTestData.scala
@@ -11,7 +11,6 @@ import edu.ie3.datamodel.models.voltagelevels.GermanVoltageLevelUtils
 import edu.ie3.simona.test.common.DefaultTestData
 import edu.ie3.util.quantities.PowerSystemUnits.PU
 import org.locationtech.jts.geom.{Coordinate, GeometryFactory, Point}
-import spire.math.Point
 import tech.units.indriya.quantity.Quantities
 
 import java.util.UUID